Web/ManaStaff_Web/requirements.txt
anyio==4.11.0
asgiref==3.9.1
CacheControl==0.14.3
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
Django==5.2.6
djangorestframework==3.16.1
firebase_admin==7.1.0
gcloud==0.18.3
google-api-core==2.25.1
google-auth==2.41.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jwcrypto==1.5.6
msgpack==1.1.1
oauth2client==4.1.3
packaging==25.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
PyJWT==2.10.1
pyparsing==3.2.5
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-jwt==4.1.0
requests==2.32.5
requests-toolbelt==0.10.1
rsa==4.9.1
setuptools==80.9.0
sib-api-v3-sdk==7.6.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
whitenoise==6.11.0
